Organized code structure of homePage
