+++ src/pages/home/assests/HComponentsImages/bags.png
