+++ src/pages/home/assests/featProducts.js
@@ -1,0 +1,40 @@
+
+
+import prod1 from "./featuredImage/babiesf.png";
+import prod2 from "./featuredImage/carf.png";
+import prod3 from "./featuredImage/iphone2f.png";
+import prod4 from "./featuredImage/iphonef.png";
+import prod5 from "./featuredImage/jewerlyf.png";
+import prod6 from "./featuredImage/shoef.png";
+import prod7 from "./featuredImage/umbref.png";
+
+export const feat= [
+    {
+        id:11,
+        fimage:prod1
+    },
+    {
+        id:12,
+        fimage:prod2
+    },
+    {
+        id:13,
+        fimage:prod3,
+    },
+    {
+        id:14,
+        fimage:prod4,
+    },
+    {
+        id:15,
+        fimage:prod5,
+    },
+    {
+        id:16,
+        fimage:prod6,
+    },
+    {
+        id:17,
+        fimage:prod7,
+    },
+];
