+++ src/pages/home/assests/featuredImage/jewerlyf.png
