+++ src/pages/home/assests/featuredImage/shoef.png
