+++ src/pages/home/assests/products.js
@@ -1,12 +1,23 @@
-import product1 from "./assets/products/1.png";
-import product2 from "./assets/products/2.png";
-import product3 from "./assets/products/3.png";
-import product4 from "./assets/products/4.png";
-import product5 from "./assets/products/5.png";
-import product6 from "./assets/products/6.png";
-import product7 from "./assets/products/7.png";
-import product8 from "./assets/products/8.png";
-import product9 from "./assets/products/9.png";
+
+import product1 from "./CatergoryImage/1.png";
+import product2 from "./CatergoryImage/2.png";
+import product3 from "./CatergoryImage/3.png";
+import product4 from "./CatergoryImage/4.png";
+import product5 from "./CatergoryImage/5.png";
+import product6 from "./CatergoryImage/6.png";
+import product7 from "./CatergoryImage/7.png";
+import product8 from "./CatergoryImage/8.png";
+import product9 from "./CatergoryImage/9.png";
+
+
+// import product from "./assets/products/3.png";
+// import product3 from "./assets/products/3.png";
+// import product4 from "./assets/products/4.png";
+// import product5 from "./assets/products/5.png";
+// import product6 from "./assets/products/6.png";
+// import product7 from "./assets/products/7.png";
+// import product8 from "./assets/products/8.png";
+// import product9 from "./assets/products/9.png";
 
 export const PRODUCTS = [
   {
