+++ src/pages/home/components/featureAnimation.jsx
@@ -3,11 +3,9 @@ export  const FeatAnimate  = (props) => {
     const { fimage } = props.fimg
     return ( 
         <div className="featAnimComponent">
-        <div>
-        <div className="featureAnim">
+       <div className="featureAnim">
         <img src={fimage} alt="Car" />
-        </div>
-        </div>
+        </div> 
         
             
         </div>
