+++ src/pages/home/context/shop-context.jsx
@@ -1,5 +1,6 @@
 import { createContext, useEffect, useState } from "react";
-import { PRODUCTS } from "../products";
+// import { PRODUCTS } from "../products";
+import { PRODUCTS } from "../assests/products";
 
 export const ShopContext = createContext(null);
 
@@ -55,3 +56,4 @@ export const ShopContextProvider = (props) => {
     </ShopContext.Provider>
   );
 };
+//
