--- src/pages/home/featProducts.js
@@ -1,39 +1,0 @@
-
-import prod1 from "./assets/products/featuredImage/babiesf.png";
-import prod2 from "./assets/products/featuredImage/carf.png";
-import prod3 from "./assets/products/featuredImage/iphone2f.png";
-import prod4 from "./assets/products/featuredImage/iphonef.png";
-import prod5 from "./assets/products/featuredImage/jewerlyf.png";
-import prod6 from "./assets/products/featuredImage/shoef.png";
-import prod7 from "./assets/products/featuredImage/umbref.png";
-
-export const feat= [
-    {
-        id:11,
-        fimage:prod1
-    },
-    {
-        id:12,
-        fimage:prod2
-    },
-    {
-        id:13,
-        fimage:prod3,
-    },
-    {
-        id:14,
-        fimage:prod4,
-    },
-    {
-        id:15,
-        fimage:prod5,
-    },
-    {
-        id:16,
-        fimage:prod6,
-    },
-    {
-        id:17,
-        fimage:prod7,
-    },
-];
