+++ src/pages/home/home.jsx
@@ -1,8 +1,8 @@
 import { BrowserRouter as Router, Routes, Route,Switch } from "react-router-dom";
 import { Navbar } from "./components/navbar";
-import { Shop } from "./pages/shop/shop";
-import { Contact } from "./pages/contact";
-import { Cart } from "./pages/cart/cart";
+import { Shop } from "./components/shop";
+// import { Contact } from "./pages/contact";
+// import { Cart } from "./pages/cart/cart";
 import { ShopContextProvider } from "./context/shop-context";
 import {Footer} from "./components/footer";
 import Hero  from "./components/hero";
