--- src/pages/home/pages/cart/cart.jsx
@@ -1,48 +1,0 @@
-import React, { useContext } from "react";
-import { ShopContext } from "../../context/shop-context";
-import { PRODUCTS } from "../../products";
-import { CartItem } from "./cart-item";
-import { useHistory } from "react-router-dom"; // Use useHistory instead of useNavigate
-
-import "./cart.css";
-
-export const Cart = () => {
-  const { cartItems, getTotalCartAmount, checkout } = useContext(ShopContext);
-  const totalAmount = getTotalCartAmount();
-
-  const history = useHistory(); // Use useHistory instead of useNavigate
-
-  return (
-    <div className="cart">
-      <div>
-        <h1>Your Cart Items</h1>
-      </div>
-      <div className="cart">
-        {PRODUCTS.map((product) => {
-          if (cartItems[product.id] !== 0) {
-            return <CartItem data={product} />;
-          }
-          return null; // Always include a return statement in the map function
-        })}
-      </div>
-
-      {totalAmount > 0 ? (
-        <div className="checkout">
-          <p> Subtotal: ${totalAmount} </p>
-          <button onClick={() => history.push("/")}> Continue Shopping </button> {/* Use history.push("/") */}
-          <button
-            onClick={() => {
-              checkout();
-              history.push("/checkout"); // Use history.push("/checkout")
-            }}
-          >
-            {" "}
-            Checkout{" "}
-          </button>
-        </div>
-      ) : (
-        <h1> Your Shopping Cart is Empty</h1>
-      )}
-    </div>
-  );
-};
