--- src/pages/home/pages/contact.jsx
@@ -1,5 +1,0 @@
-import React from "react";
-
-export const Contact = () => {
-  return <div>contact</div>;
-};
